Added views for rescheduling/list for files and menu
Also added scroll views for Archive view and Later view, and the top
segment menu switches between the scroll views. Swipe up to refresh
(and add a new message if one was acted upon before). Archived and
rescheduled messages are added to the respective scroll views. Added
shake to undo, which also removes a message from the respective scroll
views if the last action was to reschedule/archive.